// gabinet-kosmetyczny-frontend/src/app/feature/account/model/worker-health-card.interface.ts
export interface IWorkerHealthCard {
    uuid?: string;
    pacemaker?: boolean;
    hermophilia?: boolean;
    psoriasis?: boolean;
    allergies?: string;
    discoloration?: boolean;
    infectiousDiseases?: boolean;
    bloodCirculationDisorders?: boolean;
    herpes?: boolean;
    fever?: boolean;
    pregnancy?: boolean;
    weakness?: boolean;
    needsUpdate?: boolean;
}